IB/srp: Maintain a single connection per I_T nexus
The sysfs attribute 'add_target' may not be used for multiple logins to
the same target. If doing so with multipathing, this crashes the
multipath-tools. Furthermore, we want to prevent the possibility of data
corruption here. If manual reconnect is necessary, then the user can use
the 'delete' sysfs attribute of the remote port before connecting.

Note: This is different from the implementation by Bart Van Assche.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>