IB/srp: Configure blk_rq_tmo through rport sysfs
We introduce the sysfs file 'blk_rq_tmo' to make the block layer
request queue timeout configurable. We set that value directly
for all connected SCSI devices on that remote port.
We take care that we set the same timeout in srp_slave_configure()
if a new LUN is scanned afterwards.

With SRP there are only two possibilities to detect a stale
connection: QP timeout and/or block layer timeout.
Please note: The QP timeout is much more reliable. So the blk
layer timeout should only be triggered if a LUN is blocked while
the QP isn't in timeout.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>