IB/srp: Allow fast IO failing through rport sysfs
We do things differently than Bart Van Assche here. We use that
'fast_io_fail_tmo' file to set the block device timeout directly
for all SCSI devices on that remote port. We take care that we
don't race with LUN scanning and set the same timeout in
srp_slave_configure() if a new LUN is scanned afterwards.

With SRP there are only two possibilities to detect a stale
connection: QP timeout and/or block layer timeout.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>