IB/srp: Reconnect timed out paths when connecting
The SCST target tends to keep timed out connections. If a path
fails, we cannot connect it anymore without reboot.
So try to send a DREQ - even if not being connected. If this is
successful, we can try to establish the SRP connection again.

In SCST before 2012-08-06 this can trigger a race condition on the
target.
SCST commit: Fix race between session removal and session creation

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>