IB/srp: Fail IO faster based on block and QP timeout
The QP timeout is hard to calculate -> it's more like guessing.
The default is at approx. 35 seconds at 7 QP retries. As we can't
do IO when the QP timed out, the best solution would be to register
on both: See what triggers first and then fail IO.

Note: It takes 5 seconds longer than the block layer timeout as we
register this on the srp_abort() function for now.

We set all connected SCSI LUNs and the target to 'transport-offline'.
There may be no LUN scanning in parallel. We block the SCSI error
handling in srp_reset_device() until the IO failing has finished.
Then we can return success there to leave the SCSI error handling
and to be able to fail IO fast.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>