IB/srp: Fail IO faster based on QP timeout
The QP timeout is hard to calculate -> it's more like guessing.
The default is at approx. 35 seconds at 7 QP retries. But the QP
timeout is also the most reliable indication for a connection issue.
So register fast IO failing on the QP timeout.

We set all connected SCSI LUNs and the target to 'transport-offline'.
In case we've hit the block layer timeout or other issues before, we
are in SCSI error handling in parallel.
We block the SCSI error handling in srp_abort() and srp_reset_device()
until the IO failing has finished. Then we can return FAST_IO_FAIL
there to leave the SCSI error handling and to be able to fail IO fast.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>