IB/srp: Show that a reconnect can work
EXPERIMENTAL! TEMPORARY!

We add the rport sysfs file 'reconnect' in order to manually
reconnect the target. This only works if the SCSI devices
are already offline.

For this we set the target state to 'SRP_TARGET_FAILED' again
and call srp_reconnect_target(). Depending on the result we
can set the target state to 'SRP_TARGET_LIVE' or
'SRP_TARGET_RECON'.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>