IB/srp: Make the QP timeout configurable through sysfs
We use the number of QP retries in order to make the overall QP
timeout configurable. A default for it can be set through sysfs
which is taken each time a connection is established. Unfortunately,
the QP retry count can't be changed while already being connected.
It may not be less than 2 or greater than 7. If greater than 7,
there would be an overflow (8 -> 0).

The default in the code is still at 7 QP retries which is approx.
35 seconds overall QP timeout. The minimum 2 QP retries is at
approx. 14 seconds.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>