IB/srp: Fix a possible race in srp_reset_host()
It is possible that the reconnect in srp_reset_host() races with
the fast IO failing triggered by the QP timeout. Therefore, we set
the target state in srp_reset_host() also into SRP_TARGET_FAILED
before triggering the reconnect. If this doesn't work, we wait for
the fast IO failing to be finished and return.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>