IB/srp: Don't set the QP into error state in tl_err_work
It is possible that not all requests can be claimed before failing
IO in tl_err_work. In this case it is very dangerous to set the QP
into error state. The functions srp_abort() and srp_reset_device()
are the functions which run in parallel.
The main trigger for this parallel code execution is congestion.

Messages like this one were visible in the kernel log:
AMD-Vi: Event logged [IO_PAGE_FAULT device=03:00.0 ...]

Tests with setting the QP into error state without failing IO
also resulted in a kernel panic before. So never set the QP
into error state if IO is possibly still in flight!

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>