IB/srp: Set sdevs to running upon successful reconnect
We've noticed an issue in production that some LUNs were offline
after successful reconnect. This can be caused if the tl_err_work
triggers in parallel to the SCSI error handling.

A message like this can be found in the kernel log:
sd 8:0:0:1: Device offlined - not ready after error recovery

The function scsi_target_unblock() doesn't set SCSI devices in
SDEV_OFFLINE back to SDEV_RUNNING. So do that explicitly upon
successful reconnect.

Signed-off-by: Sebastian Riemer <sebastian.riemer@profitbricks.com>